+++ apps/dapp/src/components/ssov-new/Tables/StrikesChain/index.tsx
