+++ apps/dapp/src/hooks/ssov/query.ts
