+++ apps/dapp/src/hooks/ssov/useContractData.ts
