+++ apps/dapp/src/hooks/ssov/useVaultStore.ts
